Black Magic and Coffee
I did a lot of TDD until everything started working, then I realized that I hadn't made a commit in about an hour. So much has changed, but it is finally starting to work.

The next step is to have the form/subscriber model auto-populate if the Subscriber name is recognized on the server.

After that, the server should keep a record of everything that has been published. Additionally, the server should use the subscriber model to generate a record of all the notifications that it may be interested in (Whether they were logged in at the time or not.)

Finally,  the client should pull down all of those notifications if the subscriber name is recognized and display them.